Fix overzealous optimization in differentiable vcall
Enoki's differentiable vcall implementation relied on a
gather<Permute=true>() optimization that turned this operation into a
scatter under reverse-mode differentiation. This is problematic when
other parts of the AD graph accumulate gradients into the same node,
because they are potentially overwritten by the scatter. This commit
fixes the problem using one level of indirection, by scattering into a
copy and then propagating those gradients to the original node.