Fixed a tricky reentrancy bug related to AD/CustomOp
Deallocating edges with a CustomOp invokes the CustomOp destructor,
which can cause a flurry of other activity in the AD backend (e.g. scene
graph teardown). When this happens in the middle of an edge deletion
operation, it can lead to invalid intermediate state (e.g. two vertices
that are only connected with a forward but not a backward edge) that
breaks subsequent traversal operations (in my case, an infinite loop).
FWIW This was caught by assertions in debug mode.

This commit introduces a cleanup list that postpones these destructor
calls until the graph structure is known to be consistent.