Texture: keep track of migration status using a separate flag
As of recently, the `ek::replace_grad()` methods broadcasts its input
operands so that they have the same size. This caused a regression in
the Texture class, which analyzed the size of the `m_value` array to
determine whether the texture has been migrated to the CPU. This commit
adds a separate flag to keep track of this information.