Added ek.isolate_grad() context manager providing control over postponing of AD operations
For some time now, Enoki has postponed reverse-mode differentiation
steps that would exit the megakernel currently being recorded. This is
super-important because we need to run these steps later and outside of
the restricted execution environment of a megakernel. However, the
details of this process were kind of magic and coupled to details of the
megakernel recording process.

This magic is now gone, and there is a `isolate_grad()` context manager
that represents this functionality:

```
with ek.isolate_grad():
    # code using ek.backward() / ek.backward_from/to() / ek.traverse() ..
```

It can be used in any situation and stops AD traversal of edges that
cross the isolation boundary. These differentiation steps are then
performed when exiting the `with` statement. Note: `ek.Loop` and
vectorized virtual function calls implicitly open an isolation scope
(I think we always want this).

This change will allow us to run differentiable rendering algorithms
that locally use AD using *wavefront* mode as well (previously, this
only worked in megakernel mode). Besides benchmarking, this can be
useful when one wants to debug such algorithms, e.g., by printing the
contents of variables.