Fix matrix stride computation
This commit fixes an issue noticed by Kate Salesin: the strides computed
for the Matrix41f type used in monochromatic polarized simulations
within Mitsuba were incorrect, causing faulty NumPy<->Dr.Jit
conversions.

This fix is not particular elegant but does the trick. A better solution
should eventually be possible with the nanobind-based rewrite of Dr.Jit
(which uses a different technique for dr.ravel and along with a
DLPack-based interface).