Update Dr.Jit-Core, work around Clang 14 miscompilation (issue #282)
On Clang 14, `ArrayBase::shuffle_<..>()` "forgets" to call the default
constructor of the `Derived out` values. On AD variants, this causes a
reference counting failure when `out.entry(..) = ` invokes the
destructor of a field with uninitialized memory. Very strange.

This commit works around the issue by setting
`out = dr::zeros<Derived>()`. It would be good to investigate this
further and potentially submit a Clang bug at some point.