Simplified intrinsic include declarations
The file ``packet_intrin.h`` tried to include a subset of the large set
of AVX512 intrinsics. That made it dependent on implementation details
of C++ compilers/libraries and was a source of fragility. This commit
reverts to a simpler include directive with the downside of pulling in
many more headers.