C++ array ``operator<<`` and ``to_string`` rewrite
This commit makes array stringification consistent with the Python
``ArrayBase.__repr__()``  changes from the previous commit.

The main difference between the previous and current implementations is
that the former iterated over array indices sweeping from the outside
towards inside components. In a vectorized (SIMD/JIT) array, the
vectorized dimension was changing the slowest, which caused the
desirable behavior that each matrix/vector/quaternion/.. was printed
after the other. This worked for all types including matrices stored in
column major order. But following the transition to row-major matrices,
their output was now shown as transposed.

The new code checks if a type is vectorized -- if so, it iterates over
the vector dimension first, and then prints the rest sweeping from
inside to outside.

There is also a bunch of unrelated type trait renaming to make the C++
and Python interfaces consistent:

- ``array_depth_v`` -> ``depth_v``
- ``array_size_v`` -> ``size_v`