Added class ``drjit.detail.VariableTracker`` for tracking variable state
Various parts of Dr.Jit rely on the ability to repeatedly read and write
sets of Python variables to evolve them through multi-step operations
such as symbolic/evaluated loops, conditionals, etc.

Doing this right is actually surprisingly complex. We need to carefully
check that this set of variables hasn't somehow changed in structure
when we revisit it. Another aspect is that the user may update some
variables in-place while replacing other variables with completely
different ones (that's what an ``a = b``-style assignment does by
default). Depending on which form is chosen, the user will expect
different semantics from the higher-level operation.

The class ``drjit.detail.VariableTracker`` is meant to fix limitations
where Dr.Jit currently does not exhibit the expected behavior. It also
abstracts these complex details into a self-contained part of the
codebase and is meant to replace the mixture of traversal mechanisms
used in various other operations.