Switch ``dr.while_loop()`` to new ``drjit.detail.VariableTracker`` API
This commit adds a new internal API ``drjit.detail.VariableTracker``
to traverse a Python object graph multiple times and read/write variable
indices.  The class validates the consistency of this object graph over
time and raises actionable error messages if problems are found.

This functionality is needed by operations such as ``dr.while_loop()``
and ``dr.if_stmt()`` and is meant to replace several half-baked
implementations with a unified solution.  This first commit switches
``dr.while_loop()`` over to the new interface.

One big advantage of the new implementation is that it tracks when
a variable has been replaced with a different value, and when the
change was performed using in-place mutation.

This better captures regular Python semantics. For example, it would be
surprising if calling the following ``f(x)`` would mutate the variable
``x`` provided by the caller.

```python
@dr.syntax
def f(x):
    while x < 10:
        x = x + 1
```

However, such a mutation is expected in the following snippet if ``x``
is a mutable type:

```python
@dr.syntax
def f(x):
    while x < 10:
        x += 1
```

Previously, we could not distinguish between these two cases and
conservatively chose the former, which users found confusing.