Kernel freezing: frontend and unit tests
This is not intended to work as-is, but rather to inform the kind of cases we need to check for, and how to do the Python function --> frozen function inputs / outputs mapping while being aware of aliasing, literal, etc.
A fair part of the code is there to detect and prevent unsupported cases.