Digital Differential Analyzer (DDA) and analytic integration routines
This big commit consists of two parts:

The first adds a "digital differential analyzer" (DDA), which is an
algorithm that walks along the intersection of a ray segment and a
N-dimensional grid, performing a custom computation at every grid cell.
The implementation uses modular floating point arithmetic, which ensures
numerical stability even when traversing very large grids.

The second part adds the function ``drjit.integrate()``, which builds on
the DDA to compute differentiable line integrals of bi- or trilinearly
interpolants stored on a grid.