Sphinx: switch to ``enum_tools.autoenum``
This commit integrates ``enum_tools``
(https://enum-tools.readthedocs.io/en/latest/api/autoenum.html) which
automatically traverses enumerations and merges their docstrings into
the generated Sphinx reference. This finally works because the latest
version of nanobind changed the implementation of enumerations to create
subclasses of the built-in ``enum.Enum`` and ``enum.IntEnum`` types.

In the context of Dr.Jit, it means that lots of duplicated help text
can finally be removed.