Remove `ArraBase` inheritance to `ArrayBaseT`
The issue is that Empty Base class optimization is prohibited if one of the empty base classes is also the type or the base of the type of the first non-static data member.

In practice, this means that, for instance, Matrix types were no longer tightly packed, which is particularly apparent with scalar variants.

With respect to the Python bindings, this change has no impact as our array types are bound to explicitly inherit from the Python array base (see bind.cpp)