Adjust TensorFlow presence check
After uninstalling TensorFlow, I sometimes end up with a completely
empty ``tensorflow`` package on my machine. Based on the presence of
this package, Dr.Jit thinks that TensorFlow is available and then runs
many (failing) tests. This makes the import criterion a bit stricter by
attempting to import a submodule that is always present.