Properly compile ``rsqrt`` intrinsic on LLVM
Dr.Jit provides an approximate ``dr.rsqrt()`` operation that performs a
reciprocal square root with a very small amount of rounding error (~1
ULP) when compared to the IEEE 754-compliant version. The main benefit
is that this has significantly lower latency.

Previously, only the CUDA backend implemented this operation properly,
while an IEEE 754-compliant fallback was used on the LLVM backend. This
commit adds a proper implementation for AVX, AVX512, and ARM Neon.