Evaluated loop derivative improvement
This commit improves handling of evaluated loops with grad-enabled state
variables. Previously, the AD variable ID of each differentiable state
variable changed in every iteration, even if the loop did not touch that
variable. This is an implementation detail of the loop evaluation code,
that should, however, not leak into user code. This commit fixes this
behavior.