Hash grid encoding (#390)
This PR implements a HashGrid encoding layer for the upcoming neural network
interface. It works with the `nn.Sequential` module and can take `CoopVec`,
`ArrayXf` or lists of `Float` values in both 16 and 32 bit. The output will
be either a `CoopVec` of 16 or 32 bit floats depending on the storage type.

* Added PyHashGrid3D implementation

* Ported hashgrid to drjit upstream

* Added permutohedral hashgrid

Co-authored-by: tszirr <tobias.zirr@alphanew.net>

---------

Co-authored-by: Merlin Nimier-David <merlinND@users.noreply.github.com>
Co-authored-by: tszirr <tobias.zirr@alphanew.net>