``dr.rng()``: added ``symbolic`` argument
The RNG class previously tried to auto-detect when it is used in a
symbolic context. Otherwise, it maintained an opaque internal sampler
state.

This strategy is OK. However, sometimes it can be useful to work with a
purely symbolic sampler so that the underlying state constants are
*intentionally* merged into the generated program. This commit adds a
``dr.rng(.., symbolic=True)`` flag to enable this behavior.