``tp_init_array()``: improve initialization from external frameworks
The current code path that creates an array from an external
DLPack-compatible framework queries ``meta_get(arg).ndim`` (which calls
``ndarray_import()`` internally). The tensor is then imported twice,
which breaks imports of raw DLPack capsules as the capsule is marked
as consumed following the first import. Apart from this, it also seems
kind of inefficient.

This commit switches to a more careful approach to extract the number of
dimensions. For raw capsules, we directly jump to the main case without
broadcasting.