Added ``drjit.replace_shape_t()`` convenience type trait
In generic Dr.Jit functions (i.e., ones that are supposed to work with
many different inputs), it can be rather tedious to obtain a specific
intermediate or output type.

For example, perhaps a function is called with ``Float`` but then it
needs access to an equivalent `Matrix2f` type.

This commit adds a function that makes this possible--in this case,
we can run

```
Matrix2f = dr.replace_shape_t(Float, (2, 2, -1), 'matrix')
```