Green context API
This commit adds an API to create CUDA green contexts, which split the
GPU into smaller partitions with a specific number of SMs. Dr.Jit can
furthermore launch its own kernels into this green context. There is
a simple Python API:

```python
from drjit.cuda{.ad} import green_context

with green_context(16):
    ....
```

It is also possible to extract additional information, such as the
requested vs actual number of SMs (CUDA may not always be able to
satisfy the request perfectly) and the CUDA context of the remaining
set of SMs, which can be passed to a separate computation that
wishes to launch kernels there that will be isolated.

```python
from drjit.cuda import green_context

with green_context(16) as ctx:
    print(ctx.sm_count)
    print(ctx.requested_sm_count)
    print(ctx.remaining_ctx) # Capsule containing the CUDA context
```

It is valid to create a green_context once and leave/enter it multiple
times via the Python context manager.

The green_context class and its properties (sm_count, requested_sm_count,
remaining_ctx) are fully documented in the reference documentation.